+++ src/stmmodel.cpp
@@ -266,7 +266,7 @@ void STMModel::setNumSolutes(int numSolutes)
 #ifdef USE_OPENMP
 #pragma omp parallel for
 #endif
-    for(size_t i = 0 ; i < m_elements.size()  ; i++)
+    for(int i = 0 ; i < (int)m_elements.size()  ; i++)
     {
       Element *element = m_elements[i];
       element->initializeSolutes();
@@ -275,7 +275,7 @@ void STMModel::setNumSolutes(int numSolutes)
 #ifdef USE_OPENMP
 #pragma omp parallel for
 #endif
-    for(size_t i = 0 ; i < m_elementJunctions.size()  ; i++)
+    for(int i = 0 ; i < (int)m_elementJunctions.size()  ; i++)
     {
       ElementJunction *elementJunction = m_elementJunctions[i];
       elementJunction->initializeSolutes();
@@ -496,7 +496,7 @@ bool STMModel::initializeElements(std::list<string> &errors)
 #ifdef USE_OPENMP
 #pragma omp parallel for
 #endif
-  for(size_t i = 0 ; i < m_elementJunctions.size()  ; i++)
+  for(int i = 0 ; i < (int)m_elementJunctions.size()  ; i++)
   {
     ElementJunction *elementJunction = m_elementJunctions[i];
     elementJunction->index = i;
@@ -522,7 +522,7 @@ bool STMModel::initializeElements(std::list<string> &errors)
 #ifdef USE_OPENMP
 #pragma omp parallel for
 #endif
-  for(size_t i = 0 ; i < m_elements.size()  ; i++)
+  for(int i = 0 ; i < (int)m_elements.size()  ; i++)
   {
     Element *element = m_elements[i];
     element->index = i;
